#36 [Update] 로그아웃 방식 변경
Access Token을 가지고 로그아웃 하는 방식에서 Access Token 없이 로그아웃 하는 방식으로 변경

+++ src/main/java/com/mpnp/baechelin/user/service/UserService.java
@@ -1,5 +1,7 @@
 package com.mpnp.baechelin.user.service;
 
+import com.mpnp.baechelin.exception.CustomException;
+import com.mpnp.baechelin.exception.ErrorCode;
 import com.mpnp.baechelin.login.jwt.AuthToken;
 import com.mpnp.baechelin.login.jwt.AuthTokenProvider;
 import com.mpnp.baechelin.login.jwt.repository.UserRefreshTokenRepository;
@@ -11,6 +13,7 @@ import com.mpnp.baechelin.util.HeaderUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.transaction.Transactional;
@@ -23,12 +26,20 @@ public class UserService {
     private final UserRefreshTokenRepository userRefreshTokenRepository;
     private final AuthTokenProvider tokenProvider;
 
-    public void logout(HttpServletRequest request, HttpServletResponse response, String socialId) {
+    public void logout(HttpServletRequest request, HttpServletResponse response) {
+        String refreshToken = CookieUtil.getCookie(request, "refresh_token")
+                .map(Cookie::getValue)
+                .orElse((null));
+
+        if (refreshToken == null) {
+            throw new CustomException(ErrorCode.REFRESH_TOKEN_NOT_EXIST);
+        }
+
         // Cookie에 담겨있는 refresh token 삭제
         CookieUtil.deleteCookie(request, response, "refresh_token");
 
         // DB에 저장되어 있는 refresh token 삭제
-        userRefreshTokenRepository.deleteBySocialId(socialId);
+        userRefreshTokenRepository.deleteByRefreshToken(refreshToken);
     }
 
     public UserResponseDto getUserInfo(String socialId) {
